perf: isolate route data and reuse CI compiler cache
