+++ .github/workflows/ci.yml
@@ -25,6 +25,8 @@ jobs:
     name: static
     runs-on: ubuntu-24.04
     timeout-minutes: 10
+    env:
+      PUPPETEER_SKIP_DOWNLOAD: "true"
     steps:
       - name: Checkout
         uses: actions/checkout@d23441a48e516b6c34aea4fa41551a30e30af803 # v6
@@ -101,6 +103,8 @@ jobs:
     name: node
     runs-on: ubuntu-24.04
     timeout-minutes: 15
+    env:
+      PUPPETEER_SKIP_DOWNLOAD: "true"
     steps:
       - name: Checkout
         uses: actions/checkout@d23441a48e516b6c34aea4fa41551a30e30af803 # v6
@@ -130,6 +134,7 @@ jobs:
       # The offline guard (DVNS_OFFLINE_GUARD) is set per-step below,
       # AFTER dependency installation, so pip can reach PyPI.
       PYTHONPATH: scripts/etl:scripts/ci
+      PUPPETEER_SKIP_DOWNLOAD: "true"
     steps:
       - name: Checkout
         uses: actions/checkout@d23441a48e516b6c34aea4fa41551a30e30af803 # v6
@@ -150,6 +155,8 @@ jobs:
         uses: actions/setup-python@5fda3b95a4ea91299a34e894583c3862153e4b97 # v7
         with:
           python-version: "3.12"
+          cache: pip
+          cache-dependency-path: requirements-etl.txt
 
       - name: Install pinned ETL dependencies
         run: >-
@@ -207,6 +214,16 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
+      # Cache only compiler state. Runtime fetch responses must remain fresh
+      # for the production gates, and every run still builds and tests its tree.
+      - name: Restore Next compiler cache
+        uses: actions/cache@caa296126883cff596d87d8935842f9db880ef25 # v5
+        with:
+          path: .next/cache/turbopack
+          key: ${{ runner.os }}-${{ runner.arch }}-next-v1-${{ hashFiles('package-lock.json', '.nvmrc', 'next.config.ts') }}-${{ github.sha }}
+          restore-keys: |
+            ${{ runner.os }}-${{ runner.arch }}-next-v1-${{ hashFiles('package-lock.json', '.nvmrc', 'next.config.ts') }}-
+
       - name: Build
         run: npm run build
 
@@ -236,7 +253,10 @@ jobs:
         uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7
         with:
           name: lighthouse-irpef
-          path: .lighthouseci/
+          path: |
+            .lighthouseci/*.json
+            .lighthouseci/*.html
+          include-hidden-files: true
           if-no-files-found: ignore
           retention-days: 14
